Extract functions from PanZoom to utils
